requirements.txt
Flask
PyMuPDF
Pillow
werkzeug